Refactor deck module into utils directory
Moved the deck module into the utils directory for better organization and improved module accessibility across the project.
